rewrite pr with universal scheduler
