+++ src/main/java/com/syntaxphoenix/spigot/smoothtimber/compatibility/towny/TownyChopListener.java
@@ -4,6 +4,7 @@ import org.bukkit.Location;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
 
+import com.syntaxphoenix.spigot.smoothtimber.SmoothTimber;
 import com.syntaxphoenix.spigot.smoothtimber.event.AsyncPlayerChopTreeEvent;
 import com.syntaxphoenix.spigot.smoothtimber.event.reason.DefaultReason;
 
@@ -18,11 +19,13 @@ public class TownyChopListener implements Listener {
     @EventHandler
     public void onAsyncPlayerChopTree(final AsyncPlayerChopTreeEvent event) {
         for (final Location location : event.getBlockLocations()) {
-            if (!towny.canDestroy(event.getPlayer(), location.getBlock())) {
-                event.setCancelled(true);
-                event.setReason(DefaultReason.TOWNY);
-                break;
-            }
+            SmoothTimber.getScheduler().runTask(location, () -> {
+              if (!towny.canDestroy(event.getPlayer(), location.getBlock())) {
+                  event.setCancelled(true);
+                  event.setReason(DefaultReason.TOWNY);
+                  return;
+              }
+            });
         }
     }
